json serializable dataclasses + refactor dataclasses :droplet: :ok_hand:
Removed the "tableau1, tableau2,..." and the same with "foundation1..."...
We only use the "tableaus" and "foundations" lists anyway, since that is much easier to iterate through!

Also now use the "dataclasses_json" package, to easily allow going to and from json for the dataclasses Klondike and Card.
This will help translating klondike games from the client to the server!
:ok_hand: