Fixed extraction of cards from image! :ok_hand: :droplet: :droplet: :ok_hand:
New new implementation of create stacks which takes the detections and their coordinates from the original image and splits it into the three piles of cards that the game needs; pile, foundations, and tableaus :ok_hand: :sweat_drops: :ok_hand:

Added a function that takes the output of the new new create stacks function and turns it into lists of strings matching the names of the cards that was detected by the neural network. :ok_hand:

smol change to `infer_from_image` function, it now returns its detections sorted by their y-position, such that the cards are in the correct order of there are more than one

`find_number_suit` finally renamed to `find_slices`, as its purpose is to find the slices containing cards from the original image.
It also now returns both the box it uses to find the slices which contain the coordinates of each and the slice in a tuple, allowing us to get the location of the detections so we know which cards go where! :ok_hand:

Also: smol added type annotations n' stuff :ok_hand: