Flask route for getting next move from client :ok_hand: :droplet:
Added a `get_next_klondike_move` function with the `@app.route` decorator, which takes the image in json form and converts it to an image we can work with and then returns a `jsonify`'ed version of the human-readable move to the client

It does this via a helper function that allows us to test the functionality of the function without having to send requests through flask.
This is the `get_move_from_img` function. It simply takes the image, extracts the different piles from that image and then creates a Klondike object with those values, and lastly returns the `find_move_wrapper` function that finds the next move in a Klondike game object and returns a human-readable version of that move