Smol Card refactor + find_move_wrapper fix :ok_hand: :droplet:
Card refactor:
- `translate` function has been renamed to `__str__`, as it pretty much served the purpose of translating an object to a string readable by a human for the client side :ok_hand:

- Added a few type annotations...

`find_move_wrapper` fix:
- New function: `get_card_at`.
This takes a list of cards and a desired index and attempts to return that card. If there is no card, it returns None.
By default, this returns the `str(card)` since that is what was needed the most, but this can be turned off by setting the parameter `to_str` to `False`.
This function makes the code in `find_move_wrapper` MUCH more readable!

- Fixed `find_move_wrapper` function!
It now sets the respose to the output of the `get_card_at` function, removing the need for nested if this and that checks.
- Fixed a bug where we accidentally got the wrong card in the case where we are moving from one tableau to another.
- Bugfix: throwing an error when attempting to get card from an empty pile was fixed by the use of the `get_card_at` function :ok_hand:

- Also, smol example update and added type annotation for the response just for my own sanity :sweat_smile: