Update test_yolodetector.py
Fixed expected values of test.
We tried to make it so that it wouldn't break when the ML got better... But we didn't account for the change of the cutoff on the image that specifies where piles/foundations are versus where tableaus are...

So now we simply look at all the cards detected instead of each pile individually.
That fixed it! :ok_hand: