Updated move finder
Move finder has been changed to no longer return an error code.
This is to make sure that the algorithm works better when used with the client side, since most checks we usually do before returning the draw code are not feasible with a game made from the images.
Therefore, we decided to remove the error code and simply return draw when no other option is found